If card new and graded as easy, skip first repetition.
This is not really that useful, but more than the previous trick. If
the new card was graded as easy, it's probably not really new to the
user, and this could induce artificially to grade the card as easy at the
first repetition too if it occurs too soon. Mnemosyne does something
along these lines too.